Fixes error in sys test

+++ test/python/test_cusnarks_circom.py
@@ -100,8 +100,10 @@ def test_cusnarks():
       pkfile_bin=INPUT_PKBIN_F
       pkfile_zkey=INPUT_PKZKEY_F
       
-      os.remove(circuits_folder+pkfile_bin)
-      os.remove(circuits_folder+INPUT_VK_F)
+      if os.path.exists(circuits_folder+pkfile_bin):
+         os.remove(circuits_folder+pkfile_bin)
+      if os.path.exists(circuits_folder+INPUT_VK_F):
+         os.remove(circuits_folder+INPUT_VK_F)
 
       print("Launching Setup .....\n")
       GS = GrothSetup(in_circuit_f = circuits_folder+INPUT_R1CS_F, out_pk_f=circuits_folder+pkfile_bin,
@@ -110,8 +112,10 @@ def test_cusnarks():
       GS.setup()
       del GS
       
-      os.remove(circuits_folder+OUTPUT_PD_F)
-      os.remove(circuits_folder+OUTPUT_P_F)
+      if os.path.exists(circuits_folder+OUTPUT_PD_F):
+         os.remove(circuits_folder+OUTPUT_PD_F)
+      if os.path.exists(circuits_folder+OUTPUT_P_F):
+         os.remove(circuits_folder+OUTPUT_P_F)
      
       GP = GrothProver(circuits_folder+pkfile_bin, verification_key_f = circuits_folder+INPUT_VK_F, 
                        start_server=0, keep_f=circuits_folder, snarkjs=snarkjs_folder, n_gpus=4, seed=123)
@@ -120,8 +124,10 @@ def test_cusnarks():
 
       test_printResults(result, "TS : Cusnarks R1CS.  Proof: PKBIN + WTNS")
      
-      os.remove(circuits_folder+OUTPUT_PD_F)
-      os.remove(circuits_folder+OUTPUT_P_F)
+      if os.path.exists(circuits_folder+OUTPUT_PD_F):
+         os.remove(circuits_folder+OUTPUT_PD_F)
+      if os.path.exists(circuits_folder+OUTPUT_P_F):
+         os.remove(circuits_folder+OUTPUT_P_F)
       GP = GrothProver(circuits_folder+pkfile_bin, verification_key_f = circuits_folder+INPUT_VK_F, 
                        start_server=0, keep_f=circuits_folder, snarkjs=snarkjs_folder, n_gpus=4, seed=123)
       result = GP.proof(witness_wshm_f, circuits_folder+OUTPUT_P_F, circuits_folder+OUTPUT_PD_F, verify_en=1)
@@ -130,8 +136,10 @@ def test_cusnarks():
       
       test_printResults(result, "TS : Cusnarks R1CS.  Proof: PKBIN + WSHM")
       
-      os.remove(circuits_folder+OUTPUT_PD_F)
-      os.remove(circuits_folder+OUTPUT_P_F)
+      if os.path.exists(circuits_folder+OUTPUT_PD_F):
+         os.remove(circuits_folder+OUTPUT_PD_F)
+      if os.path.exists(circuits_folder+OUTPUT_P_F):
+         os.remove(circuits_folder+OUTPUT_P_F)
       GP = GrothProver(circuits_folder+pkfile_zkey, 
                        start_server=0, keep_f=circuits_folder, snarkjs=snarkjs_folder, n_gpus=4, seed=123)
       result = GP.proof(witness_wtns_f, circuits_folder+OUTPUT_P_F, circuits_folder+OUTPUT_PD_F, verify_en=1)
@@ -139,8 +147,10 @@ def test_cusnarks():
       del GP
       test_printResults(result, "TS : SNARKJS. Proof:  ZKEY + WTNS")
       
-      os.remove(circuits_folder+OUTPUT_PD_F)
-      os.remove(circuits_folder+OUTPUT_P_F)
+      if os.path.exists(circuits_folder+OUTPUT_PD_F):
+         os.remove(circuits_folder+OUTPUT_PD_F)
+      if os.path.exists(circuits_folder+OUTPUT_P_F):
+         os.remove(circuits_folder+OUTPUT_P_F)
       GP = GrothProver(circuits_folder+pkfile_zkey, 
                        start_server=0, keep_f=circuits_folder, snarkjs=snarkjs_folder, n_gpus=4, seed=123)
       result = GP.proof(witness_wshm_f, circuits_folder+OUTPUT_P_F, circuits_folder+OUTPUT_PD_F, verify_en=1)
